Correctly merge Uniprot data where gene identifiers have changed.
Refseq is currently reannotating many of their prokaryotic ids but other gene sources have not yet caught up
http://www.ncbi.nlm.nih.gov/refseq/about/prokaryotes/reannotation/
If this is the case, then the GFF file contains an old_locus_tag entry.
This commit puts this entry into Gene.secondaryIdentifier.
It then makes the uniprot load always add the gene identifier to secondaryIdentifier as well as whatever is configured to enable the merge of sources
This restores the go annotation numbers to what we would expect